글쓰기 > DB 저장 확인

+++ src/main/java/com/example/comento/repository/BoardRepository.java
@@ -1,7 +1,9 @@
 package com.example.comento.repository;
 
-import com.example.comento.domain.Board;
+import com.example.comento.entity.Board;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BoardRepository extends JpaRepository<Board, Integer> {
+import java.util.List;
+
+public interface BoardRepository extends JpaRepository<Board, Long> {
 }
